Split liquidity tests
